Rename package of compose ui elements.
