+++ app/src/main/java/com/hfad/palamarchuksuperapp/data/services/OpenAIApiHandler.kt
@@ -22,7 +22,6 @@ import io.ktor.client.request.header
 import io.ktor.client.request.post
 import io.ktor.client.request.setBody
 import io.ktor.http.ContentType
-import io.ktor.http.HttpStatusCode
 import io.ktor.http.contentType
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.coroutines.flow.MutableStateFlow
