+++ app/src/main/java/com/hfad/palamarchuksuperapp/ui/compose/StoreScreen.kt
@@ -117,9 +117,7 @@ import coil.compose.AsyncImage
 import coil.compose.SubcomposeAsyncImage
 import coil.request.ImageRequest
 import com.chargemap.compose.numberpicker.NumberPicker
-import com.example.compose.md_theme_my_royal
 import com.hfad.palamarchuksuperapp.R
-import com.hfad.palamarchuksuperapp.appComponent
 import com.hfad.palamarchuksuperapp.data.database.DATABASE_MAIN_ENTITY_PRODUCT
 import com.hfad.palamarchuksuperapp.data.database.StoreDatabase
 import com.hfad.palamarchuksuperapp.data.dtos.ProductDTO
@@ -128,6 +126,7 @@ import com.hfad.palamarchuksuperapp.data.repository.FakeStoreApiRepository
 import com.hfad.palamarchuksuperapp.data.repository.StoreRepositoryImpl
 import com.hfad.palamarchuksuperapp.domain.models.AppError
 import com.hfad.palamarchuksuperapp.domain.models.Product
+import com.hfad.palamarchuksuperapp.ui.compose.theme.md_theme_my_royal
 import com.hfad.palamarchuksuperapp.ui.compose.utils.BottomNavBar
 import com.hfad.palamarchuksuperapp.ui.compose.utils.DrawerWrapper
 import com.hfad.palamarchuksuperapp.ui.compose.utils.MyNavigationDrawer
