+++ app/src/main/java/com/hfad/palamarchuksuperapp/ui/compose/theme/Color.kt
@@ -1,4 +1,5 @@
-package com.example.compose
+package com.hfad.palamarchuksuperapp.ui.compose.theme
+
 import androidx.compose.ui.graphics.Color
 
 val md_theme_light_primary = Color(0xFF68548E)
