+++ app/src/main/java/com/hfad/palamarchuksuperapp/ui/reusable/elements/AppText.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.compose.AppTheme
+import com.hfad.palamarchuksuperapp.ui.compose.theme.AppTheme
 
 /**
  * Базовый компонент для отображения текста с поддержкой конфигурации.
